Prefer HTTPBIN_ENV_ as more specific env var prefix
This will help reinforce that only the special env vars are
reported by /env endpoint. This is especially important in
case of deploying the service to Kubernetes which by default
sets lots of service name-prefixed variables which may be
unintentionally listed by /env if the service is named httpbin
which in fact is a common default.

Signed-off-by: Mateusz Łoskot <mateusz@loskot.net>