feat: allow specifying content-type for `/base64` responses (#198)
Given that the `/base64/{data}` (and `/base64/decode/{data}`) endpoint
allows decoding arbitrary data, I think it makes sense to allow the
client to also specify the `Content-Type` of that data while maintaining
`text/plain` as the safe default.

See conversation starting here[1]
for context and motivation for this change.

[1] https://github.com/mccutchen/go-httpbin/issues/141#issuecomment-2692924359