fix(compat): `/range` supports duration parameter (#203)
I started to take a run at the `/range` incompatibilities identified in
https://github.com/mccutchen/go-httpbin/issues/172 but in the course of
working on this I have decided, for now, that perfect compatibility with
the original httpbin implementation is not worth the effort.

So here we do (kinda) support the `?duration` parameter, but it
effectively acts more as a delay before the entire response is written.
We also add drive-by tests to ensure that we're properly support
requests for multiple ranges (which the original httpbin _does not_
correctly support).

Properly supporting both `?chunk_size` and `?duration` parameters would
require fully re-implementing the stdlib's built-in support for range
requests, which would get complicated where splitting the response into
chunks spread over a given duration intersects with constructing proper
multipart responses for multiple range requests.

For now, I'm happy to leave the implementation _mostly_ as-is.