fix: /response-headers does not need escaping by default (#208)
The fix in commit 0decfd1 for a potential XSS vulnerability[1] in the
`/response-headers` endpoint made an unintentionally breaking change, by
HTML-escaping the body of the response when no explicit `Content-Type` is
specified in the incoming request.

We do not need to escape by default, because we default to a safe JSON content
type in that case.

[1]: https://github.com/mccutchen/go-httpbin/security/advisories/GHSA-528q-4pgm-wvg2